refactor(ui): hoist provider-card actions; drop duplicate SecretType
Two cleanups on top of the registry refactor.

(1) `useProviderActions()` hook collapses the 5-hook + first-agent
navigation boilerplate that every provider Card was duplicating into
one shared call:

    const actions = useProviderActions();
    // actions.remove(id, "Remove …?", "Remove")
    // actions.create(input)   // navigates if first agent
    // actions.update(input)

Each Card drops to ~25 lines of pure provider-specific glue. The
"if first agent, go to list" onboarding navigation lives once in the
hook instead of being recopied per provider.

(2) Drops UI's local `SecretType` definition. The union now lives
once in `api-server-api`; UI re-exports it through `ui/src/types.ts`.
We hit a real bug from the duplication during the OpenAI commit
(the api-server-api union was widened, the UI's wasn't, and tsc
caught it only because a downstream import happened to use it).
With one source of truth that drift is impossible.

Verified: `mise run check` clean, `mise run test` all green
(agent-runtime 69, ui 45, controller all, api-server 251). Behavior
unchanged.
