+++ packages/ui/src/modules/settings/components/openai/card.tsx
@@ -1,40 +1,21 @@
-import { useStore } from "../../../../store.js";
 import { PROVIDERS, type SecretView } from "../../../../types.js";
-import { useAgents } from "../../../agents/api/queries.js";
-import {
-  useCreateSecret,
-  useDeleteSecret,
-  useUpdateSecret,
-} from "../../../secrets/api/mutations.js";
+import { useProviderActions } from "../use-provider-actions.js";
 import { OpenAIConnected } from "./connected.js";
 import { OpenAIForm } from "./form.js";
 
 const NAME = PROVIDERS.openai.displayName;
 
-/**
- * Self-contained card for the OpenAI preset. Single-token form;
- * the registry's `pathPattern: "/v1/*"` is applied server-side at
- * create time.
- */
+/** Self-contained card for the OpenAI preset. The registry's
+ *  `pathPattern: "/v1/*"` is applied server-side at create time. */
 export function OpenAICard({ secret }: { secret?: SecretView }) {
-  const showConfirm = useStore((s) => s.showConfirm);
-  const setView = useStore((s) => s.setView);
-  const { data: agents = [] } = useAgents();
-  const createSecret = useCreateSecret();
-  const updateSecret = useUpdateSecret();
-  const deleteSecret = useDeleteSecret();
+  const actions = useProviderActions();
 
   if (secret) {
     return (
       <OpenAIConnected
         secret={secret}
-        onRemove={async () => {
-          if (!(await showConfirm(`Remove ${NAME} API key?`, "Remove Key"))) return;
-          deleteSecret.mutate({ id: secret.id });
-        }}
-        onSave={async ({ value }) => {
-          await updateSecret.mutateAsync({ id: secret.id, value });
-        }}
+        onRemove={() => actions.remove(secret.id, `Remove ${NAME} API key?`, "Remove Key")}
+        onSave={({ value }) => actions.update({ id: secret.id, value })}
       />
     );
   }
@@ -42,11 +23,7 @@ export function OpenAICard({ secret }: { secret?: SecretView }) {
   return (
     <OpenAIForm
       variant="wizard"
-      onSave={async ({ value }) => {
-        const isFirst = agents.length === 0;
-        await createSecret.mutateAsync({ type: "openai", name: NAME, value });
-        if (isFirst) setView("list");
-      }}
+      onSave={({ value }) => actions.create({ type: "openai", name: NAME, value })}
     />
   );
 }
